+++ src/main/java/com/tallerwebi/dominio/ServicioJugador.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface ServicioJugador {
     void guardarJugador(Jugador jugador) throws JugadorExistente, JugadorInexistente;
-    void eliminarJugador(Integer jugadorId) throws JugadorInexistente;
-    Boolean actualizarJugador(Integer jugadorId) throws JugadorInexistente;
-    Jugador buscarJugador(Integer jugadorId) throws JugadorInexistente;
+    void eliminarJugador(Long jugadorId) throws JugadorInexistente;
+    Boolean actualizarJugador(Long jugadorId) throws JugadorInexistente;
+    Jugador buscarJugador(Long jugadorId) throws JugadorInexistente;
 
     List<Jugador> obtenerJugadores();
 }
